+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+-   [examples]: bar chart with ACGT sequence logo
 -   [packages/core]: support for prop `interactive` in `Legend`
 -   [packages/core]: support for threshold color scales
 -   [packages/core]: support for fill and stroke animation in `Rectangle`